# perf: improve Kili client initialization time