--- a/tests/e2e/test_client.py
+++ b/tests/e2e/test_client.py
@@ -6,6 +6,13 @@ def test_client_init_not_too_long():
     timer = timeit.Timer("_ = Kili()", setup="from kili.client import Kili")
     time_spent = timer.timeit(number=1)
 
+    assert time_spent < 2
+
+
+def test_import_and_init_time_not_too_long():
+    timer = timeit.Timer("from kili.client import Kili; _ = Kili()")
+    time_spent = timer.timeit(number=1)
+
     assert time_spent < 3
 
 
@@ -26,18 +33,3 @@ def test_client_with_checks_disabled_always_faster_than_with_checks_enabled():
         time_spent_without_checks = timer_without_checks.timeit(number=1)
 
     assert time_spent_with_checks > 10 * time_spent_without_checks  # one order of magnitude
-
-
-def test_import_and_init_time_not_too_long():
-    timer = timeit.Timer("from kili.client import Kili; _ = Kili()")
-    time_spent = timer.timeit(number=1)
-
-    assert time_spent < 3
-
-
-@mock.patch.dict("os.environ", {"KILI_SDK_SKIP_CHECKS": "True"})
-def test_import_and_init_time_not_too_long_with_checks_disabled():
-    timer = timeit.Timer("from kili.client import Kili; _ = Kili()")
-    time_spent = timer.timeit(number=1)
-
-    assert time_spent < 0.5