done:
- stop storing user id and email in kiliauth saves 100ms
- add an env var `KILI_SDK_SKIP_CHECKS` that disables checks and graphql schema related stuff

`from kili.client import Kili` takes around ~700ms (mostly because of pandas)

`kili = Kili()` takes ~0s when checks are disabled

to consider:
- pandas import takes ~500ms...: It's pretty easy to add lazy loading of import. But I haven't found yet how to type pd.Dataframe 🤔  We can keep it as is for now. Maybe updating to more recent versions of pandas could improve this (we are using old version now because of python3.7)