@theodu 2023-04-26
> > for pandas, can't you do like this ?
> > ```
> > def label_as_df:
> >     import pandas as pd
> >     ...
> >     return pd.dataFrame(labels)
> > ```
> 
> yes that's the lazy importing I was talking about. But how do I add return type hints to kili.assets()? `-> Union[Iterable[Dict], pd.DataFrame]:`

Ah yeah I get it. first, we should add to do not support pandas anymore in the next major release.
I tried to remove the return type of the last assets overload method and it worked. Vscode still infer generator or list depending of the as_generator argument but we don't use pd.dataFrame anymore before to import it